py-director/requirements.txt
anyio==4.0.0
arrow==1.2.3
async-timeout==4.0.3
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
claude-api==1.0.17
click==8.1.7
colorama==0.4.6
curl-cffi==0.5.7
dnspython==2.4.2
exceptiongroup==1.1.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
loguru==0.7.0
pycparser==2.21
pymongo==4.5.0
pytchat==0.5.5
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.0
requests==2.31.0
rq==1.15.1
rq-win @ git+https://github.com/michaelbrooks/rq-win.git@c6e65e38f8a6ca99a0c8e0dd6fe35f9b92837cca # It is important for Windows
six==1.16.0
sniffio==1.3.0
times==0.7
toml==0.10.2
urllib3==2.0.4
win32-setctime==1.1.0
